Pi/requirements.txt
captcha==0.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
idna==3.10
numpy==1.24.4
packaging==24.1
pillow==10.4.0
requests==2.32.3
scikit-build==0.18.1
tflite-runtime==2.13.0
tomli==2.0.2
urllib3==2.2.3